need to use doubles for time when simulating long events.
git-svn-id: http://code.icecube.wisc.edu/svn/projects/simclasses/trunk@110011 16731396-06f5-0310-8873-f7f720988828

+++ public/simclasses/I3MCPE.h
@@ -37,7 +37,7 @@ struct I3MCPE {
   /**
    * Creation time of PE (photon arrival time)
    */ 
-  float time;
+  double time;
 
   /**
    * Number of PEs this object represents.
